requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.2
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cryptography==41.0.4
cycler==0.12.1
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.104.0
fonttools==4.54.1
frozenlist==1.4.1
greenlet==3.1.1
grpcio==1.66.2
grpcio-tools==1.66.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kafka-python==2.0.2
kiwisolver==1.4.7
langchain==0.3.3
langchain-core==0.3.10
langchain-text-splitters==0.3.0
langsmith==0.1.132
MarkupSafe==3.0.1
matplotlib==3.8.1
multidict==6.1.0
numpy==1.26.4
openai==1.16.1
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.18.0
portalocker==2.10.1
propcache==0.2.0
protobuf==5.28.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMuPDF==1.24.11
pyparsing==3.1.4
PyPDF2==3.0.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.11.3
# Editable install with no version control (reactive-ai==0.1.0)
-e /Users/ktanti/Projects/reactive-ai
redis==5.0.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.27.0
tenacity==8.2.1
tiktoken==0.8.0
tomli==2.0.2
tqdm==4.66.5
twilio==9.3.3
typing_extensions==4.8.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.23.2
Werkzeug==3.0.4
wikipedia==1.4.0
yarl==1.14.0
markdown
google-cloud
